Moved types to separate module
