+++ src/types.rs
@@ -1,0 +1,105 @@
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// A struct to store time as it is represented in a METAR
+pub struct Time {
+    /// The date the METAR was made
+    pub date: u8,
+    /// The hour the METAR was made
+    pub hour: u8,
+    /// The minute the METAR was made
+    pub minute: u8,
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// A struct representing the wind speed
+pub enum WindSpeed {
+    /// A wind speed measured in knots
+    Knot(u32),
+    /// A wind speed measured in metres per second
+    MetresPerSecond(u32),
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// A representation of wind direction
+pub enum WindDirection {
+    /// A heading defining wind direction
+    Heading(u32),
+    /// Wind direction is variable
+    Variable,
+    /// Wind speed is above 49mps or 99kt
+    Above,
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// Horizontal visibility
+pub enum Visibility {
+    /// Visibility is less than this number of metres
+    LessThanMetres(u32),
+    /// Visibility is less than this number of statute miles
+    LessThanStatuteMiles(u32),
+    /// Visibility in metres
+    Metres(u32),
+    /// Visibility in statute miles
+    StatuteMiles(u32),
+    /// Clouds and Visibility OK (CAVOK)
+    CavOK,
+    /// No significant clouds (NSC)
+    NoSignificantClouds,
+    /// Sky clear, no clouds (SKC or CLR)
+    SkyClear,
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// Measured air pressure
+pub enum Pressure {
+    /// Pressure in hectopascals
+    Hectopascals(u32),
+    /// Pressure in inches of mercury (inHg)
+    InchesMercury(u32),
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// Vertical visibility measurement
+pub enum VertVisibility {
+    /// A distance of vertical visibility
+    Distance(u32),
+    /// The vertical visibility value is present, so is reduced, but by an amount that hasn't or
+    /// cannot be measured
+    ReducedByUnknownAmount,
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// Cloud cover
+pub enum CloudLayer {
+    /// Few clouds (1/8)
+    Few(CloudType, u32),
+    /// Scattered cloud cover (3/8)
+    Scattered(CloudType, u32),
+    /// Broken cloud cover (5/8)
+    Broken(CloudType, u32),
+    /// Overcast cloud cover (7/8)
+    Overcast(CloudType, u32),
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// A cloud type description
+pub enum CloudType {
+    /// A normal cloud
+    Normal,
+    /// A cumulonimbus cloud
+    Cumulonimbus,
+    /// A towering cumulus cloud
+    ToweringCumulus,
+}
+
+#[derive(PartialEq, Eq, Clone, Debug)]
+/// Wind information
+pub struct Wind {
+    /// The wind direction, in degrees
+    pub dir: WindDirection,
+    /// The current wind speed
+    pub speed: WindSpeed,
+    /// The direction the wind may be varying between, smaller always comes first
+    pub varying: Option<(u32, u32)>,
+    /// The gusting speed of the wind
+    pub gusting: Option<WindSpeed>,
+}
